Use double number format from AuroraLib when displaying XP to display decimals properly

+++ src/main/java/gg/auroramc/levels/menu/PaginatedLevelMenu.java
@@ -123,9 +123,9 @@ public class PaginatedLevelMenu {
             if (iteratorLevel - 1 == level) {
                 var currentXP = leveler.getUserData(player).getCurrentXP();
                 var requiredXP = leveler.getRequiredXpForLevelUp(player);
-                placeholders.add(Placeholder.of("{current}", AuroraAPI.formatNumber(((Double) currentXP).longValue())));
+                placeholders.add(Placeholder.of("{current}", AuroraAPI.formatNumber(currentXP)));
                 placeholders.add(Placeholder.of("{current_short}", AuroraAPI.formatNumberShort(currentXP)));
-                placeholders.add(Placeholder.of("{required}", AuroraAPI.formatNumber(((Double) requiredXP).longValue())));
+                placeholders.add(Placeholder.of("{required}", AuroraAPI.formatNumber(requiredXP)));
                 placeholders.add(Placeholder.of("{required_short}", AuroraAPI.formatNumberShort(requiredXP)));
 
                 var bar = menuConfig.getProgressBar();
